+++ app/Models/NotificationPreference.php
@@ -8,6 +8,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 /**
  * @property int $user_id
  * @property bool $comment_replies
+ * @property bool $content_mentions
  * @property bool $space_moderation
  * @property-read User $user
  */
@@ -15,6 +16,7 @@ class NotificationPreference extends Model
 {
     protected $fillable = [
         'comment_replies',
+        'content_mentions',
         'space_moderation',
     ];
 
@@ -23,6 +25,7 @@ class NotificationPreference extends Model
     {
         return [
             'comment_replies' => 'boolean',
+            'content_mentions' => 'boolean',
             'space_moderation' => 'boolean',
         ];
     }